requirements.txt
asgiref==3.8.1
Backtesting==0.3.3
beautifulsoup4==4.12.3
bokeh==3.6.2
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozendict==2.4.6
html5lib==1.1
idna==3.10
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.2.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tornado==6.4.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
xyzservices==2024.9.0
yfinance==0.2.51
